+++ q2_types/feature_data/_format.py
@@ -1,0 +1,86 @@
+# ----------------------------------------------------------------------------
+# Copyright (c) 2016--, QIIME development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# ----------------------------------------------------------------------------
+
+import skbio.io
+import qiime.plugin.model as model
+
+
+class TaxonomyFormat(model.TextFileFormat):
+    def sniff(self):
+        with self.open() as fh:
+            # validate up to 10 data lines (i.e. non-comment lines)
+            # recreate behavior of zip(fh, range(10)) accounting for comments
+            count = 0
+            while count < 10:
+                line = fh.readline()
+                if line == '':
+                    break
+                elif line.startswith('#'):
+                    continue
+                else:
+                    cells = line.split('\t')
+                    if len(cells) < 2:
+                        return False
+                    count += 1
+            return False if count == 0 else True
+
+
+TaxonomyDirectoryFormat = model.SingleFileDirectoryFormat(
+    'TaxonomyDirectoryFormat', 'taxonomy.tsv', TaxonomyFormat)
+
+
+class DNAFASTAFormat(model.TextFileFormat):
+    def sniff(self):
+        filepath = str(self)
+        sniffer = skbio.io.io_registry.get_sniffer('fasta')
+        if sniffer(filepath)[0]:
+            generator = skbio.io.read(filepath, constructor=skbio.DNA,
+                                      format='fasta', verify=False)
+            try:
+                for seq, _ in zip(generator, range(5)):
+                    pass
+                return True
+            # ValueError raised by skbio if there are invalid DNA chars.
+            except ValueError:
+                pass
+        return False
+
+
+DNASequencesDirectoryFormat = model.SingleFileDirectoryFormat(
+    'DNASequencesDirectoryFormat', 'dna-sequences.fasta', DNAFASTAFormat)
+
+
+class PairedDNASequencesDirectoryFormat(model.DirectoryFormat):
+    left_dna_sequences = model.File('left-dna-sequences.fasta',
+                                    format=DNAFASTAFormat)
+    right_dna_sequences = model.File('right-dna-sequences.fasta',
+                                     format=DNAFASTAFormat)
+
+
+class AlignedDNAFASTAFormat(model.TextFileFormat):
+    def sniff(self):
+        filepath = str(self)
+        sniffer = skbio.io.io_registry.get_sniffer('fasta')
+        if sniffer(filepath)[0]:
+            generator = skbio.io.read(filepath, constructor=skbio.DNA,
+                                      format='fasta', verify=False)
+            try:
+                initial_length = len(next(generator))
+                for seq, _ in zip(generator, range(4)):
+                    if len(seq) != initial_length:
+                        return False
+                return True
+            # ValueError raised by skbio if there are invalid DNA chars.
+            except (StopIteration, ValueError):
+                pass
+        return False
+
+
+AlignedDNASequencesDirectoryFormat = model.SingleFileDirectoryFormat(
+    'AlignedDNASequencesDirectoryFormat', 'aligned-dna-sequences.fasta',
+    AlignedDNAFASTAFormat)
