+++ q2_types/plugin_setup.py
@@ -30,9 +30,9 @@ plugin = qiime.plugin.Plugin(
 
 importlib.import_module('q2_types.feature_table')
 importlib.import_module('q2_types.distance_matrix')
-importlib.import_module('q2_types._tree')
+importlib.import_module('q2_types.tree')
 importlib.import_module('q2_types.ordination')
 importlib.import_module('q2_types._sample_data')
-importlib.import_module('q2_types._feature_data')
-importlib.import_module('q2_types._reference_features')
+importlib.import_module('q2_types.feature_data')
+importlib.import_module('q2_types.reference_features')
 importlib.import_module('q2_types._per_sample_sequences')
