+++ q2_types/reference_features/__init__.py
@@ -1,0 +1,16 @@
+# ----------------------------------------------------------------------------
+# Copyright (c) 2016--, QIIME development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# ----------------------------------------------------------------------------
+
+import importlib
+
+from ._format import ReferenceFeaturesDirectoryFormat
+from ._type import ReferenceFeatures, SSU
+
+__all__ = ['ReferenceFeaturesDirectoryFormat', 'ReferenceFeatures', 'SSU']
+
+importlib.import_module('q2_types.reference_features._transformer')
