+++ q2_types/tree/__init__.py
@@ -1,0 +1,17 @@
+# ----------------------------------------------------------------------------
+# Copyright (c) 2016--, QIIME development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# ----------------------------------------------------------------------------
+
+import importlib
+
+from ._format import NewickFormat, NewickDirectoryFormat
+from ._type import Phylogeny, Rooted, Unrooted
+
+__all__ = [
+    'NewickFormat', 'NewickDirectoryFormat', 'Phylogeny', 'Rooted', 'Unrooted']
+
+importlib.import_module('q2_types.tree._transformer')
